+++ Lab 2/trs/Controllers/EntryController.cs
@@ -17,11 +17,12 @@ namespace trs.Controllers
         }
 
         [HttpPost]
-        public IActionResult Index(EntryViewModel model)
+        public IActionResult Index(EntryViewModel model)//when receiving submitted entry
         {
             System.Diagnostics.Debug.WriteLine("code: " + model.entry.code + " time: " + model.entry.time + " description: " + model.entry.description);
             model.projectCodes = ActivityModel.GetCodesList();
 
+            //TO DO - Redirect
             return View(model);
         }
     }