+++ Lab 2/trs/Models/EntryModel.cs
@@ -1,4 +1,5 @@
 using Microsoft.AspNetCore.Mvc.Rendering;
+using System.ComponentModel.DataAnnotations;
 
 namespace trs.Models;
 
@@ -7,7 +8,8 @@ public class EntryModel
     public string code { get; set; }
     public int time { get; set; }
     public string description { get; set; }
-    public DateOnly date { get; set; }
+    [DataType(DataType.Date)]
+    public DateTime date { get; set; }
 }
 
 public class EntryViewModel